requirements.txt
altair==6.0.0
annotated-types==0.7.0
ansible-cmdb==1.31
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
attrs==25.4.0
autocommand==2.2.2
backports.tarfile==1.2.0
bitsandbytes==0.49.1
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
braceexpand==0.1.7
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
cheroot==11.1.2
CherryPy==18.10.0
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.23.0
confection==0.1.5
contexttimer==0.3.3
cymem==2.0.13
decorator==5.2.1
decord==0.6.0
distlib==0.4.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fairscale==0.4.4
ffmpeg-python==0.2.0
filelock==3.20.0
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
hostlist==1.4.8
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
ImageIO==2.37.2
iopath==0.1.10
ipython==8.37.0
jaraco.collections==5.2.1
jaraco.context==6.0.2
jaraco.functools==4.4.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonxs==0.6
kaggle==1.7.4.5
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.13.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
portalocker==3.2.0
portend==3.2.1
pre_commit==4.5.0
preshed==3.0.12
prompt_toolkit==3.0.52
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
sacremoses==0.1.1
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
streamlit==1.52.1
sympy==1.14.0
tempora==5.8.1
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
thinc==8.3.10
tifffile==2025.5.10
timm==0.4.12
tokenizers==0.12.1
toml==0.10.2
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.31.0
triton==3.5.1
typer-slim==0.20.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
ushlex==0.99.1
virtualenv==20.35.4
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.3
webdataset==1.0.2
webencodings==0.5.1
wrapt==2.0.1
zc.lockfile==4.0
